vif precision improvements

+++ libvmaf/src/feature/funque_vif.c
@@ -39,11 +39,11 @@ void reflect_pad(const float* src, size_t width, size_t height, int reflect, flo
 
     for (unsigned int i = 0; i != reflect; i++) {
         memcpy(&dest[(reflect - 1) * out_width - i * out_width], &dest[reflect * out_width + (i + 1) * out_width], sizeof(float) * out_width);
-        memcpy(&dest[(out_height - reflect) * out_width + i * out_width], &dest[(out_height - reflect - 1) * out_width - (i+1) * out_width], sizeof(float) * out_width);
+        memcpy(&dest[(out_height - reflect) * out_width + i * out_width], &dest[(out_height - reflect - 1) * out_width - (i + 1) * out_width], sizeof(float) * out_width);
     }
 }
 
-void integral_image_2(const float* src1, const float* src2, size_t width, size_t height, float* sum)
+void integral_image_2(const float* src1, const float* src2, size_t width, size_t height, double* sum)
 {
     for (int i = 0; i < (height + 1); ++i)
     {
@@ -52,7 +52,7 @@ void integral_image_2(const float* src1, const float* src2, size_t width, size_t
             if (i == 0 || j == 0)
                 continue;
 
-            float val = src1[(i - 1) * width + (j - 1)] * src2[(i - 1) * width + (j - 1)];
+            double val = (double)src1[(i - 1) * width + (j - 1)] * (double)src2[(i - 1) * width + (j - 1)];
 
             if (i >= 1)
             {
@@ -73,7 +73,7 @@ void integral_image_2(const float* src1, const float* src2, size_t width, size_t
     }
 }
 
-void integral_image(const float* src,  size_t width, size_t height, float* sum)
+void integral_image(const float* src, size_t width, size_t height, double* sum)
 {
     for (int i = 0; i < (height + 1); ++i)
     {
@@ -81,8 +81,8 @@ void integral_image(const float* src,  size_t width, size_t height, float* sum)
         {
             if (i == 0 || j == 0)
                 continue;
- 
-            float val = src[(i - 1) * width + (j - 1)];
+
+            double val = (double)src[(i - 1) * width + (j - 1)];
 
             if (i >= 1)
             {
@@ -98,14 +98,14 @@ void integral_image(const float* src,  size_t width, size_t height, float* sum)
                     val += sum[i * width + j - 1];
                 }
             }
-            sum[i  * (width+1) + j] = val;
+            sum[i * (width + 1) + j] = val;
         }
     }
 }
 
-void compute_metrics(const float* int_1_x, const float* int_1_y, const float* int_2_x, const float* int_2_y, const float* int_xy, size_t width, size_t height, size_t kh, size_t kw, float kNorm, float* var_x, float* var_y, float* cov_xy)
+void compute_metrics(const double* int_1_x, const double* int_1_y, const double* int_2_x, const double* int_2_y, const double* int_xy, size_t width, size_t height, size_t kh, size_t kw, double kNorm, double* var_x, double* var_y, double* cov_xy)
 {
-    float mx, my, vx, vy, cxy;
+    double mx, my, vx, vy, cxy;
 
     for (unsigned int i = 0; i < (height - kh); i++)
     {
@@ -119,14 +119,14 @@ void compute_metrics(const float* int_1_x, const float* int_1_y, const float* in
 
             cxy = ((int_xy[i * width + j] - int_xy[i * width + j + kw] - int_xy[(i + kh) * width + j] + int_xy[(i + kh) * width + j + kw]) / kNorm) - (mx * my);
 
-            var_x[i * (width - kw) + j] = vx < 0 ? 0 : vx;
-            var_y[i * (width - kw) + j] = vy < 0 ? 0 : vy;
-            cov_xy[i * (width - kw) + j] = (vx < 0 || vy < 0) ? 0 : cxy;
+            var_x[i * (width - kw) + j] = vx < 0 ? (double)0 : vx;
+            var_y[i * (width - kw) + j] = vy < 0 ? (double)0 : vy;
+            cov_xy[i * (width - kw) + j] = (vx < 0 || vy < 0) ? (double)0 : cxy;
         }
     }
 }
 
-int compute_vif_funque(const float* x, const float* y, size_t width, size_t height, double *score, double *score_num, double *score_den, int k, int stride, float sigma_nsq)
+int compute_vif_funque(const float* x, const float* y, size_t width, size_t height, double* score, double* score_num, double* score_den, int k, int stride, double sigma_nsq)
 {
     int ret = 1;
 
@@ -134,7 +134,7 @@ int compute_vif_funque(const float* x, const float* y, size_t width, size_t heig
     int kw = k;
     int k_norm = k * k;
 
-    float *x_pad, *y_pad;
+    float* x_pad, * y_pad;
 
     int x_reflect = (int)((kh - stride) / 2);
     int y_reflect = (int)((kw - stride) / 2);
@@ -147,14 +147,14 @@ int compute_vif_funque(const float* x, const float* y, size_t width, size_t heig
     size_t r_width = width + (2 * x_reflect);
     size_t r_height = height + (2 * x_reflect);
 
-    float *int_1_x, *int_1_y, *int_2_x, *int_2_y, *int_xy;
-    float *mu_x, *mu_y, *var_x, *var_y, *cov_xy;
-  
-    int_1_x = (float*)calloc((r_width + 1) * (r_height + 1), sizeof(float));
-    int_1_y = (float*)calloc((r_width + 1) * (r_height + 1), sizeof(float));
-    int_2_x = (float*)calloc((r_width + 1) * (r_height + 1), sizeof(float));
-    int_2_y = (float*)calloc((r_width + 1) * (r_height + 1), sizeof(float));
-    int_xy = (float*)calloc((r_width + 1) * (r_height + 1), sizeof(float));
+    double* int_1_x, * int_1_y, * int_2_x, * int_2_y, * int_xy;
+    double* mu_x, * mu_y, * var_x, * var_y, * cov_xy;
+
+    int_1_x = (double*)calloc((r_width + 1) * (r_height + 1), sizeof(double));
+    int_1_y = (double*)calloc((r_width + 1) * (r_height + 1), sizeof(double));
+    int_2_x = (double*)calloc((r_width + 1) * (r_height + 1), sizeof(double));
+    int_2_y = (double*)calloc((r_width + 1) * (r_height + 1), sizeof(double));
+    int_xy = (double*)calloc((r_width + 1) * (r_height + 1), sizeof(double));
 
     integral_image(x_pad, r_width, r_height, int_1_x);
     integral_image(y_pad, r_width, r_height, int_1_y);
@@ -162,18 +162,18 @@ int compute_vif_funque(const float* x, const float* y, size_t width, size_t heig
     integral_image_2(y_pad, y_pad, r_width, r_height, int_2_y);
     integral_image_2(x_pad, y_pad, r_width, r_height, int_xy);
 
-    var_x = (float*)malloc(sizeof(float) * (r_width + 1 - kw) * (r_height + 1 - kh));
-    var_y = (float*)malloc(sizeof(float) * (r_width + 1 - kw) * (r_height + 1 - kh));
-    cov_xy = (float*)malloc(sizeof(float) * (r_width + 1 - kw) * (r_height + 1 - kh));
+    var_x = (double*)malloc(sizeof(double) * (r_width + 1 - kw) * (r_height + 1 - kh));
+    var_y = (double*)malloc(sizeof(double) * (r_width + 1 - kw) * (r_height + 1 - kh));
+    cov_xy = (double*)malloc(sizeof(double) * (r_width + 1 - kw) * (r_height + 1 - kh));
 
     compute_metrics(int_1_x, int_1_y, int_2_x, int_2_y, int_xy, r_width + 1, r_height + 1, kh, kw, k_norm, var_x, var_y, cov_xy);
 
     size_t s_width = (r_width + 1) - kw;
     size_t s_height = (r_height + 1) - kh;
 
-    float* g = (float*)malloc(sizeof(float) * s_width * s_height);
-    float* sv_sq = (float*)malloc(sizeof(float) * s_width * s_height);
-    float exp = 1e-10;
+    double* g = (double*)malloc(sizeof(double) * s_width * s_height);
+    double* sv_sq = (double*)malloc(sizeof(double) * s_width * s_height);
+    double exp = 1e-10;
     int index;
 
     *score = 0;
@@ -186,33 +186,33 @@ int compute_vif_funque(const float* x, const float* y, size_t width, size_t heig
         {
             index = i * s_width + j;
 
-            g[index] =  cov_xy[index] / (var_x[index] + exp);
+            g[index] = cov_xy[index] / (var_x[index] + exp);
             sv_sq[index] = var_y[index] - g[index] * cov_xy[index];
 
             if (var_x[index] < exp)
             {
-                g[index] = 0;
+                g[index] = (double)0;
                 sv_sq[index] = var_y[index];
-                var_x[index] = 0;
+                var_x[index] = (double)0;
             }
 
             if (var_y[index] < exp)
             {
-                g[index] = 0;
-                sv_sq[index] = 0;
+                g[index] = (double)0;
+                sv_sq[index] = (double)0;
             }
 
             if (g[index] < 0)
             {
                 sv_sq[index] = var_x[index];
-                g[index] = 0;
+                g[index] = (double)0;
             }
 
             if (sv_sq[index] < exp)
                 sv_sq[index] = exp;
 
-            *score_num += log(1 + g[index] * g[index] * var_x[index] / (sv_sq[index] + sigma_nsq)) + 1e-4;
-            *score_den += log(1 + var_x[index] / sigma_nsq) + 1e-4;
+            *score_num += (log(1 + g[index] * g[index] * var_x[index] / (sv_sq[index] + sigma_nsq)) + (double)1e-4);
+            *score_den += (log(1 + var_x[index] / sigma_nsq) + (double)1e-4);
         }
     }
